Comment error lines

+++ didactic/data/cardinal/predict.py
@@ -521,7 +521,7 @@ class CardiacRepresentationPredictionWriter(BasePredictionWriter):
                 list_values = TABULAR_CAT_ATTR_LABELS[attr]
                 df = df.replace({list_values[0]: 0, list_values[1]: 1, list_values[2]: 2})
                 # change type to float 
-                df[f"{attr}_probs"] = df[f"{attr}_probs"].astype(float)
+                # df[f"{attr}_probs"] = df[f"{attr}_probs"].astype(float)
             else:
                 raise ValueError(f"Unexpected number of categories for attribute {attr}: {TABULAR_CAT_ATTR_LABELS[attr]}")
         return df
